sm8450-common: Add config for display composition color spaces
Currently we use the Display-P3 color space for display composition
across all color modes on P19. We do this to facilitate  a new version
of Adaptive color mode; this forces display to always use P3 color
calibration.

It turns out that in P19's case using P3 color calibration for sRGB
content, even when the content is accurately converted to P3, degrades
color accuracy for certain sRGB colors. To fix this, we revert back to
pre-P19 behavior, where we let SurfaceFlinger choose the most
appropriate color space for composition, for all color modes used on P19
*except* for Adaptive color mode.

Bug: 137140317
Bug: 137053654
Change-Id: Id9c414de862000c59d860776277df1cb0a719b06
Signed-off-by: pppig236 <priv@pppig236.com>